+++ packages/api-derive/src/staking/ownSlashes.ts
@@ -8,9 +8,10 @@ import type { BalanceOf, EraIndex, Perbill } from '@polkadot/types/interfaces';
 import type { ITuple } from '@polkadot/types/types';
 import type { DeriveStakerSlashes } from '../types';
 
-import { map, of, switchMap } from 'rxjs';
+import { map, of } from 'rxjs';
 
 import { memo } from '../util';
+import { erasHistoricApplyAccount } from './util';
 
 export function _ownSlashes (instanceId: string, api: ApiInterfaceRx): (accountId: Uint8Array | string, eras: EraIndex[], withActive: boolean) => Observable<DeriveStakerSlashes[]> {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -41,12 +42,4 @@ export function ownSlash (instanceId: string, api: ApiInterfaceRx): (accountId: 
   );
 }
 
-export function ownSlashes (instanceId: string, api: ApiInterfaceRx): (accountId: Uint8Array | string, withActive?: boolean) => Observable<DeriveStakerSlashes[]> {
-  return memo(instanceId, (accountId: Uint8Array | string, withActive = false): Observable<DeriveStakerSlashes[]> => {
-    return api.derive.staking.erasHistoric(withActive).pipe(
-      switchMap((eras) =>
-        api.derive.staking._ownSlashes(accountId, eras, withActive)
-      )
-    );
-  });
-}
+export const ownSlashes = erasHistoricApplyAccount('_ownSlashes');