+++ packages/types/src/metadata/PortableRegistry/util.ts
@@ -1,0 +1,7 @@
+// Copyright 2017-2021 @polkadot/types authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export function assertUnreachable (_x: never): never {
+  throw new Error("Didn't expect to get here");
+}