+++ packages/types/src/metadata/decorate/storage/index.ts
@@ -8,7 +8,7 @@ import type { Storage } from '../types';
 import { lazyMethod, lazyMethods, stringCamelCase } from '@polkadot/util';
 
 import { objectNameFirstLower } from '../util';
-import { createFunction, createKeyRaw } from './createFunction';
+import { createFunction, createKeyRaw, NO_RAW_ARGS } from './createFunction';
 import { getStorage } from './getStorage';
 import { createRuntimeFunction } from './util';
 
@@ -34,7 +34,7 @@ export function decorateStorage (registry: Registry, { pallets }: MetadataLatest
           {
             palletVersion: createRuntimeFunction(
               { method: VERSION_NAME, prefix, section },
-              createKeyRaw(registry, { method: VERSION_KEY, prefix: name.toString() }, [], [], []),
+              createKeyRaw(registry, { method: VERSION_KEY, prefix: name.toString() }, NO_RAW_ARGS),
               VERSION_DOCS
             )(registry)
           },