+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 Changes:
 
+- Generated TS Enum interfaces now expose the `type` (with allowed values)
 - Add Kusama 9130 upgrade block
+- Internal maintainability cleanups
 
 
 ## 6.11.1 Dec 5, 2021